requirements.txt
torch
torchvision
timm
opencv-python
numpy
scikit-learn
streamlit
pillow
matplotlib
